manually renames all places app was hardcoded
