+++ packages/insomnia-app/config/config.json
@@ -4,7 +4,7 @@
   "userDataFolder": "Insomnia",
   "productName": "Insomnia",
   "synopsis": "The Collaborative API Client and Design Tool",
-  "icon": "https://github.com/kong/insomnia/blob/develop/packages/insomnia-app/app/icons/icon.ico?raw=true",
+  "icon": "https://github.com/kong/insomnia/blob/develop/packages/insomnia-app/src/icons/icon.ico?raw=true",
   "changelogUrl": "https://insomnia.rest/changelog",
   "theme": "default",
   "lightTheme": "studio-light",
