+++ packages/insomnia-app/scripts/generateCACerts.ts
@@ -2,7 +2,7 @@ import fs from 'fs/promises';
 import path from 'path';
 import tls from 'tls';
 
-const filePath = path.join(__dirname, '..', 'app', 'network', 'ca_certs.ts');
+const filePath = path.join(__dirname, '..', 'src', 'network', 'ca_certs.ts');
 
 const certificates = tls.rootCertificates.join('\n');
 const fileContents = `export default \`${certificates}\`;\n`;
