+++ packages/insomnia-app/send-request/index.ts
@@ -1,1 +1,1 @@
-export { getSendRequestCallbackMemDb, getSendRequestCallback } from '../app/common/send-request';
+export { getSendRequestCallbackMemDb, getSendRequestCallback } from '../src/common/send-request';
