+++ packages/insomnia-inso/src/db/adapters/insomnia-adapter.ts
@@ -26,7 +26,7 @@ import { BaseModel } from '../models/types';
  * models.unitTest.type = 'UnitTest';
  * </pre>
  *
- * @see packages/insomnia-app/app/common/import.js
+ * @see packages/insomnia-app/src/common/import.js
  */
 
 type RawTypeKey = 'api_spec'
