+++ packages/insomnia-inso/src/db/models/types.ts
@@ -1,4 +1,4 @@
-// Simplified and extracted from insomnia-app/app/models/*
+// Simplified and extracted from insomnia-app/src/models/*
 import { Database } from '../index';
 
 export interface BaseModel {
