+++ scripts/convert-to-typescript
@@ -1,6 +1,6 @@
 #! /bin/bash
 export INSOMNIA_ROOT="$HOME/src/github.com/Kong/insomnia"
-export CODE_LOCATION="$INSOMNIA_ROOT/packages/insomnia-app/app"
+export CODE_LOCATION="$INSOMNIA_ROOT/packages/insomnia-app/src"
 
 echo "
 **/bin/*
